+++ RenderSystems/Direct3D11/src/Vao/OgreD3D11TexBufferPacked.cpp
@@ -83,7 +83,7 @@ namespace Ogre
     {
         assert( cacheIdx < 16 );
 
-        const size_t formatSize = isD3D11Structured()
+        const uint32 formatSize = isD3D11Structured()
                                       ? mBytesPerElement
                                       : PixelFormatGpuUtils::getBytesPerPixel( mPixelFormat );
 
@@ -135,7 +135,7 @@ namespace Ogre
             else if( !mCachedResourceViews[i].mResourceView )
             {
                 //We create in-order. If we hit here, the next ones are also null pointers.
-                resourceView = createResourceView( i, offset, sizeBytes );
+                resourceView = createResourceView( i, (uint32)offset, (uint32)sizeBytes );
                 break;
             }
         }
@@ -143,7 +143,7 @@ namespace Ogre
         if( !resourceView )
         {
             //If we hit here, the cache is full and couldn't find a match.
-            resourceView = createResourceView( mCurrentCacheCursor, offset, sizeBytes );
+            resourceView = createResourceView( mCurrentCacheCursor, (uint32)offset, (uint32)sizeBytes );
         }
 
         return resourceView;
